% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book-meta.R
\name{book_meta}
\alias{book_meta}
\title{Get book meta data from a Gutenberg book}
\usage{
book_meta(text)
}
\arguments{
\item{text}{A character string of lines from a Gutenberg book}
}
\value{
tibble with \code{author}, \code{title} and \code{release} columns.
}
\description{
Get book meta data from a Gutenberg book
}
\examples{
book_meta(dracula[1:20])
}
