% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_words}
\alias{count_words}
\title{Count word frequency in text}
\usage{
count_words(text)
}
\arguments{
\item{text}{A character vector}
}
\value{
A tibble with columns \code{word},
and \code{n} the number of times \code{word} occurs in \code{text}.
}
\description{
Counts the occurrence of words over all elements provided in \code{text}.
}
\details{
Each element of \code{text} is considered a line.
Lines are first converted to lower case then split into words
by whitespace. Then punctuation is removed, and empty strings are ignored.
}
\examples{
count_words(dracula)
}
