% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dracula.R
\docType{data}
\name{dracula}
\alias{dracula}
\title{Dracula}
\format{
A character vector of length 15973, each element being
a line from the Gutenberg copy of the book.
}
\source{
\url{http://www.gutenberg.org/ebooks/345}
}
\usage{
dracula
}
\description{
The text from the book Dracula by Bram Stoker
}
\keyword{datasets}
