% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_zipfs}
\alias{fit_zipfs}
\title{Estimate the Zipf's Law \eqn{\alpha} parameter}
\usage{
fit_zipfs(n)
}
\arguments{
\item{n}{A numeric vector containing word frequencies}
}
\value{
The estimate of \eqn{\alpha} as a numeric value.
}
\description{
Estimate the power law parameter, \eqn{\alpha}, for Zipf's Law from a
set of word frequencies.
}
\details{
Estimated using maximum likelihood as outlined in:
Moreno-Sánchez, Isabel, Francesc Font-Clos, and Álvaro Corral.
2016. “Large-Scale Analysis of Zipf’s Law in English Texts.”
PLoS ONE 11 (1): e0147073. \url{https://doi.org/10.1371/journal.pone.0147073.}
}
\examples{
word_freqs <- count_words(dracula)
fit_zipfs(word_freqs$n)
}
