% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_rank}
\alias{plot_rank}
\title{Rank-frequency plot}
\usage{
plot_rank(n)
}
\arguments{
\item{n}{Numeric vector of word frequencies}
}
\value{
ggplot2 plot object
}
\description{
Plots the rank vs. frequency on log-log scales, useful for
finding power law behaviour in word frequencies.
}
\examples{
word_freqs <- count_words(dracula)
plot_rank(word_freqs$n)
}
